Renamed basetextfield to amplifytextfield
