+++ packages/vue/src/components/index.ts
@@ -13,6 +13,6 @@ export { default as ConfirmSignUp } from './confirm-sign-up.vue';
 export { default as ConfirmSignIn } from './confirm-sign-in.vue';
 export { default as VerifyUser } from './verify-user.vue';
 export { default as ConfirmVerifyUser } from './confirm-verify-user.vue';
-export { default as BaseTextField } from './primitives/base-text-field.vue';
+export { default as AmplifyTextField } from './primitives/amplify-text-field.vue';
 export { default as AmplifyCheckBox } from './primitives/amplify-check-box.vue';
 export * from '../types';
